// lib/captions/types.ts
export type CaptionWordType = "word" | "punctuation" | "silence";
// ...
export interface CaptionWord {
  id?: string;
  text: string;
  type: CaptionWordType;
  /** ms from the start of the clip. */
  startTime: number;
  /** ms from the start of the clip. */
  endTime: number;
}